+++ services/distribution/src/main/java/app/coronawarn/server/services/distribution/assembly/diagnosiskeys/structure/file/TemporaryExposureKeyExportFile.java
@@ -106,8 +106,8 @@ public class TemporaryExposureKeyExportFile extends FileOnDisk {
         .setKeyData(ByteString.copyFrom(diagnosisKey.getKeyData()))
         .setTransmissionRiskLevel(diagnosisKey.getTransmissionRiskLevel())
         // TODO cwa-server/#233 Rolling start number and period should be int32
-        .setRollingStartIntervalNumber(Math.toIntExact(diagnosisKey.getRollingStartNumber()))
-        .setRollingPeriod(Math.toIntExact(diagnosisKey.getRollingPeriod()))
+        .setRollingStartIntervalNumber(diagnosisKey.getRollingStartNumber())
+        .setRollingPeriod(diagnosisKey.getRollingPeriod())
         .build())
         .collect(Collectors.toSet());
   }
